requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
Marshmallow-Peewee==3.0.0
nacos-sdk-python==0.1.1
peewee==3.13.1
psycopg2==2.8.4
pycodestyle==2.5.0
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
